authenticated and authorized user with the help of JWT

+++ controller/expenseController.js
@@ -6,7 +6,7 @@ try {
     console.log(req.body)
     const {name,amount,description} = req.body;
 
-    const expense = await Expense.create({name:name,amount:amount,description:description});
+    const expense = await Expense.create({name:name,amount:amount,description:description,userId:req.user.id});
     if(!expense) return res.status(401).json({"status":"fail","message":"something went wrong"});
 
     res.status(200).json({"status":"success","message":"successfully created"})
@@ -17,8 +17,10 @@ try {
 }
 
 module.exports.getExpense = async(req,res)=>{
+    
     try {
-        const expenses = await Expense.findAll();
+        const expenses = await Expense.findAll({where:{userId:req.user.id}});
+      
        
         if(!expenses) return res.status(404).json({"status":"fail","message":"not found"});
 
@@ -34,7 +36,7 @@ module.exports.deleteExpense = async(req,res)=>{
         const id = req.params.id
       
         if(id){
-          const response = await  Expense.destroy({where:{id:id}})
+          const response = await  Expense.destroy({where:{id:id,userId:req.user.id}})
           return res.status(200).json({"status":"success","message":"successfully deleted"})
         }
         res.status(401).json({"status":"fail","message":"id not found"})
